Add files via upload
This is a capstone project on stroke prediction dataset. Here the objective is to find the major factors contributing to stroke occurrence. The dataset consists of 11 clinical features which is significant in stroke prediction. Based on these features , I have analysed the data and plotted comparison graphs among them to  make it more clear so that we can easily find out which factors have a major contribution and how it varies(such as age, bmi, avg glucose level).  EDA, visualisation of data, data modelling are done in an effective way. Out of 5 models used,  gradient boost model  is  comparatively the best model which seemed to be suitable for the dataset.